Don't revoke token on 2.1.0 and above

+++ Buildkite/scripts/base.sh
@@ -58,8 +58,12 @@ function map_ip {
 function revoke_token {
     local token=${1}
     local orka_endpoint=${2}
+    
+    api_version=$(curl $orka_endpoint/health-check | jq ".api_version" | sed 's/[\."]//g')
 
-    echo "Revoking token..."
-    token_response=$(curl -s -m 60 -H "Content-Type: application/json" -H "Authorization: Bearer $token" -X DELETE $orka_endpoint/token)
-    echo "Token revoked: $token_response"
+    if [ $api_version -lt 210 ]; then
+        echo "Revoking token..."
+        token_response=$(curl -s -m 60 -H "Content-Type: application/json" -H "Authorization: Bearer $token" -X DELETE $orka_endpoint/token)
+        echo "Token revoked: $token_response"
+    fi
 }
